Remove unused telemetry proto (#35)
This proto is not used across multiple projects and pulling it out of
the Santa repo will create extra work each time we need to add a field.
Until we have a good reason to do so, we'll leave this where it is and
remove this extra copy to avoid confusion.