sync: add faa to the sync protocol (#38)
FAA watch items will be stored on the client in a new sqlite database.
Watch items will be sent by the sync server during the rule download
phase after all process rules have been sent and stored in their
database. A `RuleDownloadResponse` will either populate `rules` or
`file_access_watch_items`. The cursor will be used by the sync server to
handle the various stages of the rule download. Watch items will be
paginated and only sync new items.

During preflight and postflight a hash of the current watch items are
calculated and sent to the server. During preflight, if there is a
mismatch with the hash, the server will initiate a clean watch items
sync. Just like it does for rule.

In short, watch items will be handled just like rules during sync and
stored in their own database.